initial commit
Has support for pulling Aliyun bootimage data from
`openshift/installer`

Has support for creating an Aliyun client and creating a TagRequest
that we can use to mark images as bootimages.

Still needs support for:
  - parsing all the builds from a release (i.e. fetching builds.json)
  - grabbing all the Aliyun images from individual builds (parsing
    meta.json)
  - generating DeleteImageRequests
  - dry-run support

I found going through the dense `models.py` useful for helping with
the Aliyun client stuff because the docs are basically non-existent.

https://raw.githubusercontent.com/aliyun/alibabacloud-python-sdk/master/ecs-20140526/alibabacloud_ecs20140526/models.py